+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import Calendar from "./components/Calendar";
+import Calendar from "./components/Calendar/Calendar";
 import styled from "styled-components";
 
 function App() {
